Changes: - Enable -Wall and -Wextra compilation flag to improve code quality. It is the first step towards enabling -Werror. - Fixed couple dozen compilation warnings - Fixed a bug in LinearDisassemble found during this exercise.
Before (when building with "make VERBOSE=1"):
/usr/bin/c++  -Dpybinlex_EXPORTS -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/include -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/build/LIEF/include -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/build/capstone/include -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/build/tlsh/src/tlsh/include -isystem /home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/bindings/python/pybind11/include -isystem /usr/include/python3.8  -fPIC -fvisibility=hidden   -flto -fno-fat-lto-objects -o CMakeFiles/pybinlex.dir/bindings/python/blelf.cpp.o -c /home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/bindings/python/blelf.cpp

After (when building with "make VEBOSE=1", notice -Wall, -Wextra):
/usr/bin/c++  -Dbinlex_EXPORTS -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/include -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/build/LIEF/include -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/build/capstone/include -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/build/tlsh/src/tlsh/include -I/home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/bindings/python/pybind11/include -I/usr/include/python3.8  -Wall -Wextra -fPIC   -DCMAKE_CXX_FLAGS_RELEASE=-O3 -DNDEBUG -o CMakeFiles/binlex.dir/src/blelf.cpp.o -c /home/cvisinescu/projects/GeekWeek2022/binlex_staging/binlex/src/blelf.cpp

 Changes to be committed:
	modified:   CMakeLists.txt
	modified:   include/decompiler.h
	modified:   include/decompilerbase.h
	modified:   src/auto.cpp
	modified:   src/binlex.cpp
	modified:   src/blelf.cpp
	modified:   src/common.cpp
	modified:   src/decompiler.cpp
	modified:   src/decompilerbase.cpp
	modified:   src/pe.cpp

+++ src/common.cpp
@@ -99,7 +99,7 @@ string Common::GetFileSHA256(char *file_path){
 string Common::Wildcards(uint count){
     stringstream s;
     s << "";
-    for (int i = 0; i < count; i++){
+    for (uint i = 0; i < count; i++){
         s << "?? ";
     }
     return TrimRight(s.str());
@@ -174,7 +174,7 @@ string Common::WildcardTrait(string trait, string bytes){
         bytes.erase(bytes.length() - 3);
         size_t index = trait.find(bytes, 0);
         if (index != string::npos){
-            for (int j = index; j < trait.length(); j = j + 3){
+            for (size_t j = index; j < trait.length(); j = j + 3){
                 trait.replace(j, 2, "??");
             }
             break;
@@ -187,7 +187,7 @@ string Common::HexdumpBE(const void *data, size_t size){
     stringstream bytes;
     bytes << "";
     const unsigned char *local_pc = (const unsigned char *)data;
-    for (int i = 0; i < size; i++){
+    for (size_t i = 0; i < size; i++){
         bytes << hex << setfill('0') << setw(2) << (uint32_t)local_pc[i] << " ";
     }
     return TrimRight(bytes.str());
@@ -196,7 +196,7 @@ string Common::HexdumpBE(const void *data, size_t size){
 string Common::HexdumpMemDisp(uint64_t disp){
     stringstream bytes;
     const unsigned char *local_pc = (const unsigned char *)&disp;
-    for (int i = 0; i < sizeof(disp) -1 ; i++){
+    for (size_t i = 0; i < sizeof(disp) -1 ; i++){
         if (local_pc[i] != 0 && local_pc[i] != 255){
             bytes << hex << setfill('0') << setw(2) << (uint32_t)local_pc[i] << " ";
         }
